refactored to put solutions in separate file as public methods
