+++ hicap_analysis/analysis_project.py
@@ -1,5 +1,6 @@
-from hicap_analysis.wells import GPM2CFD, Well, WellResponse
+from hicap_analysis.wells import Well, WellResponse
 from hicap_analysis.utilities import Q2ts
+from hicap_analysis.solutions import GPM2CFD
 import numpy as np
 import pandas as pd
 import yaml, os, shutil
