+++ hicap_analysis/tests/data/test_ts.csv
@@ -1,0 +1,366 @@
+year,month,day,sequential_day,well1,well2,well3,well4,well5
+1,1,1,1,0,0,0,0,0
+1,1,2,2,0,0,0,0,0
+1,1,3,3,0,0,0,0,0
+1,1,4,4,0,0,0,0,0
+1,1,5,5,0,0,0,0,0
+1,1,6,6,0,0,0,0,0
+1,1,7,7,0,0,0,0,0
+1,1,8,8,0,0,0,0,0
+1,1,9,9,0,0,0,0,0
+1,1,10,10,0,0,0,0,0
+1,1,11,11,0,0,0,0,0
+1,1,12,12,0,0,0,0,0
+1,1,13,13,0,0,0,0,0
+1,1,14,14,0,0,0,0,0
+1,1,15,15,0,0,0,0,0
+1,1,16,16,0,0,0,0,0
+1,1,17,17,0,0,0,0,0
+1,1,18,18,0,0,0,0,0
+1,1,19,19,0,0,0,0,0
+1,1,20,20,0,0,0,0,0
+1,1,21,21,0,0,0,0,0
+1,1,22,22,0,0,0,0,0
+1,1,23,23,0,0,0,0,0
+1,1,24,24,0,0,0,0,0
+1,1,25,25,0,0,0,0,0
+1,1,26,26,0,0,0,0,0
+1,1,27,27,0,0,0,0,0
+1,1,28,28,0,0,0,0,0
+1,1,29,29,0,0,0,0,0
+1,1,30,30,0,0,0,0,0
+1,1,31,31,0,0,0,0,0
+1,2,1,32,0,0,0,0,0
+1,2,2,33,0,0,0,0,0
+1,2,3,34,0,0,0,0,0
+1,2,4,35,0,0,0,0,0
+1,2,5,36,0,0,0,0,0
+1,2,6,37,0,0,0,0,0
+1,2,7,38,0,0,0,0,0
+1,2,8,39,0,0,0,0,0
+1,2,9,40,0,0,0,0,0
+1,2,10,41,0,0,0,0,0
+1,2,11,42,0,0,0,0,0
+1,2,12,43,0,0,0,0,0
+1,2,13,44,0,0,0,0,0
+1,2,14,45,0,0,0,0,0
+1,2,15,46,0,0,0,0,0
+1,2,16,47,0,0,0,0,0
+1,2,17,48,0,0,0,0,0
+1,2,18,49,0,0,0,0,0
+1,2,19,50,0,0,0,0,0
+1,2,20,51,0,0,0,0,0
+1,2,21,52,0,0,0,0,0
+1,2,22,53,0,0,0,0,0
+1,2,23,54,0,0,0,0,0
+1,2,24,55,0,0,0,0,0
+1,2,25,56,0,0,0,0,0
+1,2,26,57,0,0,0,0,0
+1,2,27,58,0,0,0,0,0
+1,2,28,59,0,0,0,0,0
+1,3,1,60,0,0,0,0,0
+1,3,2,61,0,0,0,0,0
+1,3,3,62,0,0,0,0,0
+1,3,4,63,0,0,0,0,0
+1,3,5,64,0,0,0,0,0
+1,3,6,65,0,0,0,0,0
+1,3,7,66,0,0,0,0,0
+1,3,8,67,0,0,0,0,0
+1,3,9,68,0,0,0,0,0
+1,3,10,69,0,0,0,0,0
+1,3,11,70,0,0,0,0,0
+1,3,12,71,0,0,0,0,0
+1,3,13,72,0,0,0,0,0
+1,3,14,73,0,0,0,0,0
+1,3,15,74,0,0,0,0,0
+1,3,16,75,0,0,0,0,0
+1,3,17,76,0,0,0,0,0
+1,3,18,77,0,0,0,0,0
+1,3,19,78,0,0,0,0,0
+1,3,20,79,0,0,0,0,0
+1,3,21,80,0,0,0,0,0
+1,3,22,81,0,0,0,0,0
+1,3,23,82,0,0,0,0,0
+1,3,24,83,0,0,0,0,0
+1,3,25,84,0,0,0,0,0
+1,3,26,85,0,0,0,0,0
+1,3,27,86,0,0,0,0,0
+1,3,28,87,0,0,0,0,0
+1,3,29,88,0,0,0,0,0
+1,3,30,89,0,0,0,0,0
+1,3,31,90,0,0,0,0,0
+1,4,1,91,0,0,0,0,0
+1,4,2,92,0,0,0,0,0
+1,4,3,93,0,0,0,0,0
+1,4,4,94,0,0,0,0,0
+1,4,5,95,0,0,0,0,0
+1,4,6,96,0,0,0,0,0
+1,4,7,97,0,0,0,0,0
+1,4,8,98,0,0,0,0,0
+1,4,9,99,0,0,0,0,0
+1,4,10,100,0,0,0,0,0
+1,4,11,101,0,0,0,0,0
+1,4,12,102,0,0,0,0,0
+1,4,13,103,0,0,0,0,0
+1,4,14,104,0,0,0,0,0
+1,4,15,105,0,0,0,0,0
+1,4,16,106,0,0,0,0,0
+1,4,17,107,0,0,0,0,0
+1,4,18,108,0,0,0,0,0
+1,4,19,109,0,0,0,0,0
+1,4,20,110,0,0,0,0,0
+1,4,21,111,0,0,0,0,0
+1,4,22,112,0,0,0,0,0
+1,4,23,113,0,0,0,0,0
+1,4,24,114,0,0,0,0,0
+1,4,25,115,0,0,0,0,0
+1,4,26,116,0,0,0,0,0
+1,4,27,117,0,0,0,0,0
+1,4,28,118,0,0,0,0,0
+1,4,29,119,0,0,0,0,0
+1,4,30,120,0,0,0,0,0
+1,5,1,121,0,0,0,0,0
+1,5,2,122,0,0,0,0,0
+1,5,3,123,0,0,0,0,0
+1,5,4,124,0,0,0,0,0
+1,5,5,125,0,0,0,0,0
+1,5,6,126,0,0,0,0,0
+1,5,7,127,0,0,0,0,0
+1,5,8,128,0,0,0,0,0
+1,5,9,129,0,0,0,0,0
+1,5,10,130,0,0,0,0,0
+1,5,11,131,0,0,0,0,0
+1,5,12,132,0,0,0,0,0
+1,5,13,133,0,0,0,0,0
+1,5,14,134,0,0,0,0,0
+1,5,15,135,0,0,0,0,0
+1,5,16,136,0,0,0,0,0
+1,5,17,137,0,0,0,0,0
+1,5,18,138,0,0,0,0,0
+1,5,19,139,0,0,0,0,0
+1,5,20,140,0,0,0,0,0
+1,5,21,141,0,0,0,0,0
+1,5,22,142,0,0,0,0,0
+1,5,23,143,0,0,0,0,0
+1,5,24,144,0,0,0,0,0
+1,5,25,145,0,0,0,0,0
+1,5,26,146,0,0,0,0,0
+1,5,27,147,0,0,0,0,0
+1,5,28,148,0,0,0,0,0
+1,5,29,149,0,0,0,0,0
+1,5,30,150,0,0,0,0,0
+1,5,31,151,0,0,0,0,0
+1,6,1,152,0,0,0,0,0
+1,6,2,153,0,0,0,0,0
+1,6,3,154,0,0,0,0,0
+1,6,4,155,0,0,0,0,0
+1,6,5,156,0,0,0,0,0
+1,6,6,157,0,0,0,0,0
+1,6,7,158,0,0,0,0,0
+1,6,8,159,0,0,0,0,0
+1,6,9,160,0,0,0,0,0
+1,6,10,161,0,0,0,0,0
+1,6,11,162,0,0,0,0,0
+1,6,12,163,0,0,0,0,0
+1,6,13,164,0,0,0,0,0
+1,6,14,165,0,0,0,0,0
+1,6,15,166,0,0,0,0,0
+1,6,16,167,0,0,0,0,0
+1,6,17,168,0,0,0,0,0
+1,6,18,169,0,0,0,0,0
+1,6,19,170,0,0,0,0,0
+1,6,20,171,0,0,0,0,0
+1,6,21,172,0,0,0,0,0
+1,6,22,173,0,0,0,0,0
+1,6,23,174,0,0,0,0,0
+1,6,24,175,0,0,0,0,0
+1,6,25,176,0,0,0,0,0
+1,6,26,177,0,0,0,0,0
+1,6,27,178,0,0,0,0,0
+1,6,28,179,0,0,0,0,0
+1,6,29,180,0,0,0,0,0
+1,6,30,181,0,0,0,0,0
+1,7,1,182,0,0,0,0,0
+1,7,2,183,0,0,0,0,0
+1,7,3,184,0,0,0,0,0
+1,7,4,185,0,0,0,0,0
+1,7,5,186,0,0,0,0,0
+1,7,6,187,0,0,0,0,0
+1,7,7,188,0,0,0,0,0
+1,7,8,189,0,0,0,0,0
+1,7,9,190,0,0,0,0,0
+1,7,10,191,0,0,0,0,0
+1,7,11,192,0,0,0,0,0
+1,7,12,193,0,0,0,0,0
+1,7,13,194,0,0,0,0,0
+1,7,14,195,0,0,0,0,0
+1,7,15,196,0,0,0,0,0
+1,7,16,197,0,0,0,0,0
+1,7,17,198,0,0,0,0,0
+1,7,18,199,0,0,0,0,0
+1,7,19,200,0,0,0,0,0
+1,7,20,201,0,0,0,0,0
+1,7,21,202,0,0,0,0,0
+1,7,22,203,0,0,0,0,0
+1,7,23,204,0,0,0,0,0
+1,7,24,205,0,0,0,0,0
+1,7,25,206,0,0,0,0,0
+1,7,26,207,0,0,0,0,0
+1,7,27,208,0,0,0,0,0
+1,7,28,209,0,0,0,0,0
+1,7,29,210,0,0,0,0,0
+1,7,30,211,0,0,0,0,0
+1,7,31,212,0,0,0,0,0
+1,8,1,213,0,0,0,0,0
+1,8,2,214,0,0,0,0,0
+1,8,3,215,0,0,0,0,0
+1,8,4,216,0,0,0,0,0
+1,8,5,217,0,0,0,0,0
+1,8,6,218,0,0,0,0,0
+1,8,7,219,0,0,0,0,0
+1,8,8,220,0,0,0,0,0
+1,8,9,221,0,0,0,0,0
+1,8,10,222,0,0,0,0,0
+1,8,11,223,0,0,0,0,0
+1,8,12,224,0,0,0,0,0
+1,8,13,225,0,0,0,0,0
+1,8,14,226,0,0,0,0,0
+1,8,15,227,0,0,0,0,0
+1,8,16,228,0,0,0,0,0
+1,8,17,229,0,0,0,0,0
+1,8,18,230,0,0,0,0,0
+1,8,19,231,0,0,0,0,0
+1,8,20,232,0,0,0,0,0
+1,8,21,233,0,0,0,0,0
+1,8,22,234,0,0,0,0,0
+1,8,23,235,0,0,0,0,0
+1,8,24,236,0,0,0,0,0
+1,8,25,237,0,0,0,0,0
+1,8,26,238,0,0,0,0,0
+1,8,27,239,0,0,0,0,0
+1,8,28,240,0,0,0,0,0
+1,8,29,241,0,0,0,0,0
+1,8,30,242,0,0,0,0,0
+1,8,31,243,0,0,0,0,0
+1,9,1,244,0,0,0,0,0
+1,9,2,245,0,0,0,0,0
+1,9,3,246,0,0,0,0,0
+1,9,4,247,0,0,0,0,0
+1,9,5,248,0,0,0,0,0
+1,9,6,249,0,0,0,0,0
+1,9,7,250,0,0,0,0,0
+1,9,8,251,0,0,0,0,0
+1,9,9,252,0,0,0,0,0
+1,9,10,253,0,0,0,0,0
+1,9,11,254,0,0,0,0,0
+1,9,12,255,0,0,0,0,0
+1,9,13,256,0,0,0,0,0
+1,9,14,257,0,0,0,0,0
+1,9,15,258,0,0,0,0,0
+1,9,16,259,0,0,0,0,0
+1,9,17,260,0,0,0,0,0
+1,9,18,261,0,0,0,0,0
+1,9,19,262,0,0,0,0,0
+1,9,20,263,0,0,0,0,0
+1,9,21,264,0,0,0,0,0
+1,9,22,265,0,0,0,0,0
+1,9,23,266,0,0,0,0,0
+1,9,24,267,0,0,0,0,0
+1,9,25,268,0,0,0,0,0
+1,9,26,269,0,0,0,0,0
+1,9,27,270,0,0,0,0,0
+1,9,28,271,0,0,0,0,0
+1,9,29,272,0,0,0,0,0
+1,9,30,273,0,0,0,0,0
+1,10,1,274,0,0,0,0,0
+1,10,2,275,0,0,0,0,0
+1,10,3,276,0,0,0,0,0
+1,10,4,277,0,0,0,0,0
+1,10,5,278,0,0,0,0,0
+1,10,6,279,0,0,0,0,0
+1,10,7,280,0,0,0,0,0
+1,10,8,281,0,0,0,0,0
+1,10,9,282,0,0,0,0,0
+1,10,10,283,0,0,0,0,0
+1,10,11,284,0,0,0,0,0
+1,10,12,285,0,0,0,0,0
+1,10,13,286,0,0,0,0,0
+1,10,14,287,0,0,0,0,0
+1,10,15,288,0,0,0,0,0
+1,10,16,289,0,0,0,0,0
+1,10,17,290,0,0,0,0,0
+1,10,18,291,0,0,0,0,0
+1,10,19,292,0,0,0,0,0
+1,10,20,293,0,0,0,0,0
+1,10,21,294,0,0,0,0,0
+1,10,22,295,0,0,0,0,0
+1,10,23,296,0,0,0,0,0
+1,10,24,297,0,0,0,0,0
+1,10,25,298,0,0,0,0,0
+1,10,26,299,0,0,0,0,0
+1,10,27,300,0,0,0,0,0
+1,10,28,301,0,0,0,0,0
+1,10,29,302,0,0,0,0,0
+1,10,30,303,0,0,0,0,0
+1,10,31,304,0,0,0,0,0
+1,11,1,305,0,0,0,0,0
+1,11,2,306,0,0,0,0,0
+1,11,3,307,0,0,0,0,0
+1,11,4,308,0,0,0,0,0
+1,11,5,309,0,0,0,0,0
+1,11,6,310,0,0,0,0,0
+1,11,7,311,0,0,0,0,0
+1,11,8,312,0,0,0,0,0
+1,11,9,313,0,0,0,0,0
+1,11,10,314,0,0,0,0,0
+1,11,11,315,0,0,0,0,0
+1,11,12,316,0,0,0,0,0
+1,11,13,317,0,0,0,0,0
+1,11,14,318,0,0,0,0,0
+1,11,15,319,0,0,0,0,0
+1,11,16,320,0,0,0,0,0
+1,11,17,321,0,0,0,0,0
+1,11,18,322,0,0,0,0,0
+1,11,19,323,0,0,0,0,0
+1,11,20,324,0,0,0,0,0
+1,11,21,325,0,0,0,0,0
+1,11,22,326,0,0,0,0,0
+1,11,23,327,0,0,0,0,0
+1,11,24,328,0,0,0,0,0
+1,11,25,329,0,0,0,0,0
+1,11,26,330,0,0,0,0,0
+1,11,27,331,0,0,0,0,0
+1,11,28,332,0,0,0,0,0
+1,11,29,333,0,0,0,0,0
+1,11,30,334,0,0,0,0,0
+1,12,1,335,0,0,0,0,0
+1,12,2,336,0,0,0,0,0
+1,12,3,337,0,0,0,0,0
+1,12,4,338,0,0,0,0,0
+1,12,5,339,0,0,0,0,0
+1,12,6,340,0,0,0,0,0
+1,12,7,341,0,0,0,0,0
+1,12,8,342,0,0,0,0,0
+1,12,9,343,0,0,0,0,0
+1,12,10,344,0,0,0,0,0
+1,12,11,345,0,0,0,0,0
+1,12,12,346,0,0,0,0,0
+1,12,13,347,0,0,0,0,0
+1,12,14,348,0,0,0,0,0
+1,12,15,349,0,0,0,0,0
+1,12,16,350,0,0,0,0,0
+1,12,17,351,0,0,0,0,0
+1,12,18,352,0,0,0,0,0
+1,12,19,353,0,0,0,0,0
+1,12,20,354,0,0,0,0,0
+1,12,21,355,0,0,0,0,0
+1,12,22,356,0,0,0,0,0
+1,12,23,357,0,0,0,0,0
+1,12,24,358,0,0,0,0,0
+1,12,25,359,0,0,0,0,0
+1,12,26,360,0,0,0,0,0
+1,12,27,361,0,0,0,0,0
+1,12,28,362,0,0,0,0,0
+1,12,29,363,0,0,0,0,0
+1,12,30,364,0,0,0,0,0
+1,12,31,365,0,0,0,0,0
